backend/user/app/requirements.txt
fastapi
uvicorn
fastapi[all]
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
python-jose[cryptography]
bcrypt
python-dotenv
PyJWT